Fix google breadcrumbs
Remove old breadcrumbs as these were invalid and replace with new ones.

For reference:
https://developers.google.com/structured-data/breadcrumbs

Example output:

```
<div class="google-breadcrumbs is-hidden">
  <div id="google-breadcrumb-0" itemprop="child" itemref="google-breadcrumb-1" itemscope="" itemtype="http://data-vocabulary.org/Breadcrumb">
    <a href="http://www.lonelyplanet.com/thorntree" itemprop="url">
      <span itemprop="title">
        Thorn Tree
      </span>
    </a>
  </div>
  <div id="google-breadcrumb-1" itemprop="child" itemscope="" itemtype="http://data-vocabulary.org/Breadcrumb">
    <a href="http://www.lonelyplanet.com//thorntree/categories/country-forums" itemprop="url">
      <span itemprop="title">
        Country forums
      </span>
    </a>
  </div>
</div>
```

passes validation and produces valid output using:
https://developers.google.com/structured-data/testing-tool/

Does not include "Home" breadcrumb, it doesn't make much sense because
site name is going to be used by google anyways.

Does not include very last item because this is going to be a page
title and is not really a proper breadcrumb.

The final output for above example would be:
`www.lonelyplanet.com > Thorn Tree > Country forums`